requirements.txt
streamlit>=1.33.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.30.0  # Compatible with both NumPy versions
torch==2.0.1
PyPDF2==3.0.1
pandas==2.0.3
numpy>=1.26.4  # Updated for dtypes compatibility
# New libraries for multi-format support
pdfplumber==0.9.0
requests==2.31.0
beautifulsoup4==4.12.2
python-docx==0.8.11
